requirements.txt
Django==2.1.1
djangorestframework==3.9.0
geographiclib==1.49
geopy==1.17.0
pytz==2018.7
